+++ Session.h
@@ -39,8 +39,8 @@ public:
     SourceList::iterator find (std::string name);
     SourceList::iterator find (Node *node);
 
-    SourceList::iterator find (int id);
-    std::list<int> getIdList() const;
+    SourceList::iterator find (uint64_t id);
+    std::list<uint64_t> getIdList() const;
 
     SourceList::iterator at (int index);
     int index (SourceList::iterator it) const;
