+++ View.h
@@ -85,7 +85,7 @@ protected:
     virtual void saveSettings();
 
     std::string current_action_;
-    int current_id_;
+    uint64_t current_id_;
     Mode mode_;
 };
